add docs and reorganize
