+++ router_macro/Cargo.toml
@@ -16,9 +16,11 @@ proc-macro = true
 proc-macro-error2 = { version = "2.0", default-features = false }
 proc-macro2 = "1.0"
 quote = "1.0"
+syn = { version = "2.0", features = ["full"] }
 
 [dev-dependencies]
 leptos_router = { path = "../router" }
+leptos_macro = { path = "../leptos_macro" }
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(leptos_debuginfo)'] }
